+++ integration_tests/test_bergermeer.py
@@ -1,10 +1,12 @@
-from threedigrid_builder.application import make_grid
-from threedigrid_builder.constants import LineType, NodeType
-from unittest.mock import Mock
 from numpy.testing import assert_array_equal
-import numpy as np
+from threedigrid_builder.application import make_grid
+from threedigrid_builder.constants import CrossSectionShape
+from threedigrid_builder.constants import LineType
+from threedigrid_builder.constants import NodeType
+from unittest.mock import Mock
 
 import h5py
+import numpy as np
 import pathlib
 
 
@@ -67,6 +69,17 @@ def test_integration(tmp_path):
 
         ## PUMPS
         assert f["pumps"]["id"].shape == (20,)
+        assert_array_equal(f["pumps"]["id"][:], np.arange(f["pumps"]["id"].shape[0]))
+
+        ## CROSS SECTIONS
+        assert_array_equal(
+            f["cross_sections"]["id"][:], np.arange(f["cross_sections"]["id"].shape[0])
+        )
+        assert count_unique(f["cross_sections"]["shape"]) == {
+            -9999: 1,
+            CrossSectionShape.CIRCLE: 8,
+            CrossSectionShape.RECTANGLE: 3,
+        }
 
     # progress increases
     args = [x[0] for x in progress_callback.call_args_list]