+++ threedigrid_builder/base/array.py
@@ -1,4 +1,5 @@
 from enum import IntEnum
+from threedigrid_builder.exceptions import SchematisationError
 
 import numpy as np
 import sys
@@ -12,7 +13,12 @@ else:
     from typing import _Union
     from typing import TupleMeta
 
-__all__ = ["array_of", "is_tuple_type", "is_int_enum", "unpack_optional_type"]
+__all__ = [
+    "array_of",
+    "is_tuple_type",
+    "is_int_enum",
+    "unpack_optional_type",
+]
 
 
 def is_tuple_type(_type):
@@ -181,16 +187,13 @@ class ArrayDataClass:
         }
         return self.__class__(**new_fields)
 
-    def id_to_index(self, id):
+    def id_to_index(self, id, check_exists=False):
         """Find the index of records with given id.
-
-        Note:
-            It is not checked if the id actually exists! In that case
-            behaviour of numpy.searchsorted is produced: the returned index is
-            the index where to insert id to maintain order in self.id.
 
         Args:
             id (int or array_like): The id(s) to find in self.id
+            check_exists (bool): Whether to check if the id is actually present. Raises
+              SchematisationError if an id is not present.
 
         Returns:
             int or array_like: the indexes into self.id
@@ -199,7 +202,16 @@ class ArrayDataClass:
         # len(id)    len(self.id)   timing (microseconds)
         # 1000       2000           44
         # 1000       10000          62
-        return np.searchsorted(self.id, id)  # inverse of self.id[index]
+        result = np.searchsorted(self.id, id)  # inverse of self.id[index]
+        if check_exists:
+            check = self.index_to_id(result)
+            if not np.all(check == id):
+                missing = sorted(set(id) - set(check))
+                raise SchematisationError(
+                    f"Some objects refer to non-existing "
+                    f"{self.__class__.__name__} (missing: {missing})."
+                )
+        return result
 
     def index_to_id(self, index):
         """Find the id of records with given index.