+++ threedigrid_builder/grid/__init__.py
@@ -7,7 +7,8 @@ The domain layer does not depend on the interface or application layers.
 
 from .channels import *  # NOQA
 from .connection_nodes import *  # NOQA
-from .cross_sections import *  # NOQA
+from .cross_section_definitions import *  # NOQA
+from .cross_section_locations import *  # NOQA
 from .pipes import *  # NOQA
 from .grid_refinement import *  # NOQA
 from .grid import *  # NOQA