+++ threedigrid_builder/grid/channels.py
@@ -2,7 +2,7 @@ from threedigrid_builder.base import array_of
 from threedigrid_builder.constants import CalculationType
 from threedigrid_builder.constants import ContentType
 from threedigrid_builder.grid import linear
-from threedigrid_builder.grid.cross_sections import interpolate
+from threedigrid_builder.grid.cross_section_locations import interpolate
 
 import pygeos
 
@@ -39,8 +39,8 @@ class Channels(linear.BaseLinear):
         """
         # dpumax is interpolated between cross section location bank levels
         dpumax = interpolate(
-            nodes.cross1[node_idx],
-            nodes.cross2[node_idx],
+            nodes.cross_loc1[node_idx],
+            nodes.cross_loc2[node_idx],
             nodes.cross_weight[node_idx],
             locations,
             "bank_level",